Renamed database to contexts
