+++ src/contexts/context_impls/sqlite_database_context.rs
@@ -1,31 +1,34 @@
-use crate::database::context_traits::DatabaseContextTrait;
-use async_trait::async_trait;
+use crate::contexts::context_traits::DatabaseContextTrait;
 use migration::{Migrator, MigratorTrait};
+use sea_orm::prelude::async_trait::async_trait;
 use sea_orm::{ConnectionTrait, Database, DatabaseConnection, DbBackend, DbErr};
+use std::fmt::Debug;
 use std::sync::Arc;
+
 #[derive(Debug)]
-pub struct PostgresDatabaseContext {
+pub struct SQLiteDatabaseContext {
     pub(crate) db_connection: DatabaseConnection,
 }
-impl PostgresDatabaseContext {
-    pub async fn new(connection_string: &str) -> Result<PostgresDatabaseContext, DbErr> {
+
+impl SQLiteDatabaseContext {
+    pub async fn new(connection_string: &str) -> Result<SQLiteDatabaseContext, DbErr> {
         let db = Database::connect(connection_string).await?;
 
         let db = match db.get_database_backend() {
-            DbBackend::Postgres => db,
-            _ => panic!("Expected postgresql connection string"),
+            DbBackend::Sqlite => db,
+            _ => panic!("Expected sqlite connection string"),
         };
 
-        Ok(PostgresDatabaseContext { db_connection: db })
+        Ok(SQLiteDatabaseContext { db_connection: db })
     }
 }
 
 #[async_trait]
-impl DatabaseContextTrait for PostgresDatabaseContext {
+impl DatabaseContextTrait for SQLiteDatabaseContext {
     async fn reset(&self) -> Result<Arc<dyn DatabaseContextTrait>, DbErr> {
         Migrator::fresh(&self.db_connection).await.unwrap();
 
-        Ok(Arc::new(PostgresDatabaseContext {
+        Ok(Arc::new(SQLiteDatabaseContext {
             db_connection: self.get_connection(),
         }))
     }
