+++ src/contexts/context_traits/access_context_trait.rs
@@ -1,5 +1,5 @@
 use crate::api::server::server::AccessInfo;
-use crate::database::context_traits::EntityContextTrait;
+use crate::contexts::context_traits::EntityContextTrait;
 use crate::entities::access;
 use async_trait::async_trait;
 use sea_orm::DbErr;
