+++ src/contexts/context_traits/in_use_context_trait.rs
@@ -1,4 +1,4 @@
-use crate::database::context_traits::EntityContextTrait;
+use crate::contexts::context_traits::EntityContextTrait;
 use crate::entities::in_use;
 
 pub trait InUseContextTrait: EntityContextTrait<in_use::Model> {}
