+++ src/contexts/context_traits/project_context_trait.rs
@@ -1,5 +1,5 @@
 use crate::api::server::server::ProjectInfo;
-use crate::database::context_traits::EntityContextTrait;
+use crate::contexts::context_traits::EntityContextTrait;
 use crate::entities::project;
 use async_trait::async_trait;
 use sea_orm::DbErr;
