+++ src/contexts/context_traits/user_context_trait.rs
@@ -1,4 +1,4 @@
-use crate::database::context_traits::EntityContextTrait;
+use crate::contexts::context_traits::EntityContextTrait;
 use crate::entities::user;
 use async_trait::async_trait;
 use sea_orm::DbErr;
