+++ src/controllers/controller_impls/access_controller.rs
@@ -5,8 +5,8 @@ use crate::api::server::server::{
     CreateAccessRequest, DeleteAccessRequest, ListAccessInfoRequest, ListAccessInfoResponse,
     UpdateAccessRequest,
 };
+use crate::contexts::context_traits::{AccessContextTrait, UserContextTrait};
 use crate::controllers::controller_traits::AccessControllerTrait;
-use crate::database::context_traits::{AccessContextTrait, UserContextTrait};
 use crate::entities::{access, user};
 use std::sync::Arc;
 use tonic::{Code, Request, Response, Status};
@@ -69,9 +69,9 @@ impl AccessControllerTrait for AccessController {
         }
     }
 
-    /// Creates an access in the database.
+    /// Creates an access in the contexts.
     /// # Errors
-    /// Returns an error if the database context fails to create the access
+    /// Returns an error if the contexts context fails to create the access
     async fn create_access(
         &self,
         request: Request<CreateAccessRequest>,
@@ -120,7 +120,7 @@ impl AccessControllerTrait for AccessController {
     ///
     /// Returns a `Status` as response
     ///
-    /// `project_id` and `user_id` is set to 'default' since they won't be updated in the database.
+    /// `project_id` and `user_id` is set to 'default' since they won't be updated in the contexts.
     async fn update_access(
         &self,
         request: Request<UpdateAccessRequest>,
@@ -180,10 +180,10 @@ impl AccessControllerTrait for AccessController {
         }
     }
 
-    /// Deletes the an Access from the database. This has no sideeffects.
+    /// Deletes the an Access from the contexts. This has no sideeffects.
     ///
     /// # Errors
-    /// This function will return an error if the access does not exist in the database.
+    /// This function will return an error if the access does not exist in the contexts.
     async fn delete_access(
         &self,
         request: Request<DeleteAccessRequest>,
