+++ src/tests/contexts/in_use_context.rs
@@ -1,7 +1,7 @@
 use crate::tests::database::helpers::*;
 use crate::{
-    database::context_impls::InUseContext,
-    database::context_traits::EntityContextTrait,
+    contexts::context_impls::InUseContext,
+    contexts::context_traits::EntityContextTrait,
     entities::{in_use, project, session, user},
     to_active_models,
 };
