+++ src/tests/mod.rs
@@ -1,3 +1,3 @@
 pub mod api;
+pub mod contexts;
 pub mod controllers;
-pub mod database;
